Telfe 1383/change to async (#423)
* fix: change layers to be in async func

The styles seem to be getting applied to the layer only sometimes.
This suggests a race condition which I have managed to prove by
looking at the debug id's of the layer before and after the `applyStyle`
is invoked. `applyStyle` is an async function

* fix: switch out refs for state

Refs are turning out to be just too much trouble to maintain with
`ol` and `react` vying for control.

* fix(LayerSelector): switch from ref to state

* test(ensureLayers.test.ts): get tests passing again

* fix: rework maps to use states

* fix: persist selected layer

* Apply suggestions from code review

Signed-off-by: davec504 <88431810+davec504@users.noreply.github.com>

---------

Signed-off-by: davec504 <88431810+davec504@users.noreply.github.com>